Issue #3619: Migrate runs to archive table according to user/group configuration - fix for async and transactions
